add sell back lottery function and constant class

+++ posttest/src/main/java/com/kbtg/bootcamp/posttest/lottery/model/response/ViewLotteryPurchase.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public record ViewLotteryPurchase (
         List<String> tickets,
-        Integer count,
-        Integer cost
+        Long count,
+        Double cost
 ){
 }
